refactor: reorganize profile serializers and separate retrieve endpoint
- Move ProfileDetailedSerializer from posts_app to user_app for better organization
- Add profile_type field to ProfileSerializer, ProfileOptionSerializer, and ProfileDetailedSerializer
- Separate profile retrieval from update/delete operations:
  - Profile retrieval now handled by posts_app RetrieveProfileView
  - Profile update/delete handled by user_app UpdateDestroyProfileView
- Add trailing slash to profile retrieve endpoint for URL consistency
- Remove unused PetTypeSerializer import from posts_app
- Add typing.Literal and django.db.models.Q imports to support new functionality
